[REF] Combine installation scripts for RHEL8 to the main scripts [ENH] Minor optimization

+++ script/install_freesurfer.sh
@@ -1,13 +1,10 @@
 #!/bin/bash
 set -e
 
-# Get setup and script root directory
-if [ -z "${SETUP_PREFIX}" ]; then
-    echo "SETUP_PREFIX is not set or is empty. Defaulting to \${HOME}/Softwares."
-    export SETUP_PREFIX='${HOME}/Softwares'
-fi
+# Initialize environment
+source "$(cd "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)/utils.sh" && init_setup
 # Set environment variables
-INSTALL_PREFIX="$(eval "echo ${SETUP_PREFIX}/freesurfer")"
+INSTALL_PREFIX="$(eval "echo ${INSTALL_ROOT_PREFIX}/freesurfer")"
 FREESURFER_VERSION=${FREESURFER_VERSION:-7.4.1}
 
 # Cleanup old installation
@@ -15,7 +12,7 @@ if [ -d ${INSTALL_PREFIX} ]; then
     # backup license.txt from existed FreeSurfer folder
     if [ -f ${INSTALL_PREFIX}/license.txt ]; then
         echo "Backup FreeSurfer license from existed installation ..."
-        cp ${INSTALL_PREFIX}/license.txt $(eval "echo ${SETUP_PREFIX}")/license.txt
+        cp ${INSTALL_PREFIX}/license.txt $(eval "echo ${INSTALL_ROOT_PREFIX}")/license.txt
     fi
     echo "Cleanup old FreeSurfer installation ..."
     rm -rf ${INSTALL_PREFIX}
@@ -24,30 +21,41 @@ fi
 # Install
 echo "Installing FreeSurfer from offical website..."
 mkdir -p ${INSTALL_PREFIX}
+
+
+if [ "$OS_TYPE" == "macos" ]; then
 wget -q https://surfer.nmr.mgh.harvard.edu/pub/dist/freesurfer/${FREESURFER_VERSION}/freesurfer-macOS-darwin_x86_64-${FREESURFER_VERSION}.tar.gz \
     -P ${INSTALL_PREFIX}
 tar -xzf ${INSTALL_PREFIX}/freesurfer-macOS-darwin_x86_64-${FREESURFER_VERSION}.tar.gz -C ${INSTALL_PREFIX} --strip-components 2
 rm ${INSTALL_PREFIX}/freesurfer-macOS-darwin_x86_64-${FREESURFER_VERSION}.tar.gz
 
+# Put app to /Applications folder
+if [[ -d /Applications/Freeview.app || -L /Applications/Freeview.app ]]; then rm /Applications/Freeview.app; fi
+ln -s ${INSTALL_PREFIX}/Freeview.app /Applications/Freeview.app
+
+
+elif [ "$OS_TYPE" == "rhel8" ]; then
+wget -q https://surfer.nmr.mgh.harvard.edu/pub/dist/freesurfer/${FREESURFER_VERSION}/freesurfer-linux-centos8_x86_64-${FREESURFER_VERSION}.tar.gz \
+    -P ${INSTALL_PREFIX}
+tar -xzf ${INSTALL_PREFIX}/freesurfer-linux-centos8_x86_64-${FREESURFER_VERSION}.tar.gz -C ${INSTALL_PREFIX} --strip-components 2
+rm ${INSTALL_PREFIX}/freesurfer-linux-centos8_x86_64-${FREESURFER_VERSION}.tar.gz
+fi
+
 # Move previous license.txt to new FreeSurfer folder
-if [ -f $(eval "echo ${SETUP_PREFIX}")/license.txt ]; then
+if [ -f $(eval "echo ${INSTALL_ROOT_PREFIX}")/license.txt ]; then
     echo "Move previous FreeSurfer license to new installation ..."
-    mv $(eval "echo ${SETUP_PREFIX}")/license.txt ${INSTALL_PREFIX}/license.txt
+    mv $(eval "echo ${INSTALL_ROOT_PREFIX}")/license.txt ${INSTALL_PREFIX}/license.txt
 else
     echo "Use default personal FreeSurfer license ..."
     base64 --decode <<<  emhpZmFuZy55ZS5mZ2htQGdtYWlsLmNvbQozMDgyNwogKkNBanR5YkNZNDByTQogRlM5dmVNeDhnbnVxUQo= > ${INSTALL_PREFIX}/license.txt
 fi
-
-# Put app to /Applications folder
-if [[ -d /Applications/Freeview.app || -L /Applications/Freeview.app ]]; then rm /Applications/Freeview.app; fi
-ln -s ${INSTALL_PREFIX}/Freeview.app /Applications/Freeview.app
 
 # Add following lines into .zshrc
 echo "
 Add following line to .zshrc
 
 # FreeSurfer
-export FREESURFER_HOME=\"${SETUP_PREFIX}/freesurfer\"
+export FREESURFER_HOME=\"${INSTALL_ROOT_PREFIX}/freesurfer\"
 export \\
     OS=Linux \\
     FS_OVERRIDE=0 \\
