+++ script/install_fsleyes.sh
@@ -1,21 +1,14 @@
 #!/bin/bash
 set -e
 
-# Get setup and script root directory
-if [ -z "${SETUP_PREFIX}" ]; then
-    echo "SETUP_PREFIX is not set or is empty. Defaulting to \${HOME}/Softwares."
-    export SETUP_PREFIX='${HOME}/Softwares'
-fi
+# Initialize environment
+source "$(cd "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)/utils.sh" && init_setup
 # Set environment variables
-INSTALL_PREFIX="$(eval "echo ${SETUP_PREFIX}/fsleyes")"
+INSTALL_PREFIX="$(eval "echo ${INSTALL_ROOT_PREFIX}/fsleyes")"
 ENV_PREFIX=${INSTALL_PREFIX}/env
-# Check micromamba
-if ! command -v micromamba &> /dev/null; then
-    echo "Error: micromamba is not installed." >&2
-    exit 1
-fi
 
 # Cleanup old installation
+command -v micromamba &> /dev/null || { echo "Error: micromamba is not installed." >&2; exit 1; }
 if [ $(micromamba env list | grep -c ${ENV_PREFIX}) -ne 0 ]; then
     echo "Cleanup old environment ${ENV_PREFIX}..."
     micromamba env remove -p ${ENV_PREFIX} -yq
@@ -31,9 +24,11 @@ micromamba create -p ${ENV_PREFIX} -c conda-forge -yq fsleyes
 mkdir -p ${INSTALL_PREFIX}/bin
 ln -s ${ENV_PREFIX}/bin/fsleyes ${INSTALL_PREFIX}/bin/fsleyes
 
-# Put app to /Applications folder
-if [[ -d /Applications/FSLeyes.app || -L /Applications/FSLeyes.app ]]; then rm /Applications/FSLeyes.app; fi
-ln -s ${ENV_PREFIX}/share/fsleyes/FSLeyes.app /Applications/FSLeyes.app
+if [ "$OS_TYPE" == "macos" ]; then
+    # Put app to /Applications folder
+    if [[ -d /Applications/FSLeyes.app || -L /Applications/FSLeyes.app ]]; then rm /Applications/FSLeyes.app; fi
+    ln -s ${ENV_PREFIX}/share/fsleyes/FSLeyes.app /Applications/FSLeyes.app
+fi
 
 # Cleanup
 micromamba clean -yaq
@@ -43,7 +38,7 @@ echo "
 Add following lines to .zshrc:
 
 # FSLeyes
-export PATH=\"${SETUP_PREFIX}/fsleyes/bin:\${PATH}\"
+export PATH=\"${INSTALL_ROOT_PREFIX}/fsleyes/bin:\${PATH}\"
 # Note: to use this version of fsleyes
 # above lines should be put after FSL related lines
 "
